ADDED VAR COST EXCLUD FLAG
To add the var cost excluded flag I needed to add CNTRCT_TIER_ID and CUST_ID to the IPC and index those fields. Then I created a new table with distinct combos of those fields and joined out to cust elig to get the excluded flags which I'll join back in. If I try to go directly to the exclusions flag the process takes over 2hrs. this way takes 20 mins.